requirements.txt
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.3
gunicorn==19.9.0
psycopg2==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
